# openapi.yaml
openapi: 3.1.0
info:
  title: PrimeSoft Alliance Unified API
  version: 1.0.0
  description: >
    Production-grade, API-first SaaS platform for AI chatbots, bookings, contacts, forms,
    content management, multimedia, and analytics.
servers:
  - url: "https://{domain}/v1"
    description: "Production Environment (Render or custom domain)"
    variables:
      domain:
        default: "your-app.onrender.com"
        description: "The host domain of your application"
  - url: "http://localhost:3000/v1"
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /leads:
    post:
      summary: Capture a Lead
      tags: [Leads]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        '201':
          description: Lead created
    get:
      summary: List Leads
      tags: [Leads]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Paginated leads
  /leads/{lead_id}:
    get:
      summary: Lead details
      tags: [Leads]
      parameters:
        - name: lead_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Lead details
    patch:
      summary: Update lead
      tags: [Leads]
      parameters:
        - name: lead_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Lead updated
    delete:
      summary: Delete lead
      tags: [Leads]
      parameters:
        - name: lead_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Deleted
...
  /content/media:
    get:
      summary: List Media Assets
      tags: [Media]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: List Media
    post:
      summary: Register Media Asset
      tags: [Media]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mime_type, size]
              properties:
                filename: { type: string }
                mime_type: { type: string }
                size: { type: integer }
      responses:
        '201':
          description: Media registered
  /content/media/{media_id}:
    get:
      summary: Get media details
      tags: [Media]
      parameters:
        - name: media_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Media details
    patch:
      summary: Update media metadata
      tags: [Media]
      parameters:
        - name: media_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Media updated
    delete:
      summary: Delete media
      tags: [Media]
      parameters:
        - name: media_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Media deleted

  # --- Webhooks ---
...
  /analytics/leads:
    get:
      summary: Leads analytics
      tags: [Analytics]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Leads Analytics
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HeaderClientId:
      name: x-client-id
      in: header
      required: true
      schema: { type: string }
...
  schemas:
...
    Lead:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
...